requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
anthropic>=0.40.0
ccxt>=4.0.0
numpy>=1.24.0
scipy>=1.11.0
dtaidistance>=2.3.0
python-telegram-bot>=20.0
Pillow>=10.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
pandas>=2.0.0
pydantic>=2.0.0
